Added Tasks for specific Tests, added all tests to their proper classes
