+++ src/main/java/org/CandyLand/CandyLand.java
@@ -8,7 +8,7 @@ public class CandyLand {
 
     public static void main(String[] args) {
         try{
-            MainFrame mainFrame = new MainFrame();
+            MainFrame mainFrame = new MainFrame(0);
         }catch(IOException e){
             System.err.println("Background image not found.");
             System.exit(1);
